Swapped views -> init for nodes
