+++ ftplugin/k8s_nodes.lua
@@ -1,4 +1,4 @@
-local node_view = require("kubectl.views.nodes.views")
+local node_view = require("kubectl.views.nodes")
 local string_util = require("kubectl.utils.string")
 local root_view = require("kubectl.views.root.views")
 local hl = require("kubectl.actions.highlight")
