+++ lua/kubectl/init.lua
@@ -3,7 +3,7 @@ local config = require("kubectl.config")
 local pod_view = require("kubectl.views.pods")
 local deployment_view = require("kubectl.views.deployments")
 local events_view = require("kubectl.views.events")
-local nodes_view = require("kubectl.views.nodes.views")
+local nodes_view = require("kubectl.views.nodes")
 local secrets_view = require("kubectl.views.secrets.views")
 local services_view = require("kubectl.views.services.views")
 local filter_view = require("kubectl.views.filter.view")
