+++ lua/kubectl/views/nodes/init.lua
@@ -1,57 +1,25 @@
-local time = require("kubectl.utils.time")
-local find = require("kubectl.utils.find")
+local definition = require("kubectl.views.nodes.definition")
+local ResourceBuilder = require("kubectl.resourcebuilder")
+
 local M = {}
 
--- Define the custom match function for prefix and suffix
-local function match_prefix_suffix(key, _, prefix, suffix)
-  return key:match("^" .. prefix) or key:match(suffix .. "$")
+function M.Nodes()
+  ResourceBuilder:new("nodes", { "get", "nodes", "-A", "-o=json" })
+    :fetch()
+    :decodeJson()
+    :process(definition.processRow)
+    :prettyPrint(definition.getHeaders)
+    :addHints({
+      { key = "<d>", desc = "describe" },
+    }, true, true)
+    :display("k8s_nodes", "Nodes")
 end
 
-local function getRole(row)
-  local key, _ = find.dictionary(row.metadata.labels, function(key, value)
-    return match_prefix_suffix(key, value, find.escape("node-role.kubernetes.io/"), find.escape("kubernetes.io/role"))
-  end)
-
-  if key then
-    --TODO: Not sure if this handles the second kubernetes.io/role match
-    local role = vim.split(key, "/")
-    if #role == 2 then
-      return role[2]
-    end
-  end
-  return ""
-end
-
-local function getStatus(row)
-  --TODO: Get status based on conditions
-  return ""
-end
-function M.processRow(rows)
-  local data = {}
-  for _, row in pairs(rows.items) do
-    local pod = {
-      name = row.metadata.name,
-      status = getStatus(row),
-      roles = getRole(row),
-      age = time.since(row.metadata.creationTimestamp),
-      version = row.status.nodeInfo.kubeletVersion,
-    }
-
-    table.insert(data, pod)
-  end
-  return data
-end
-
-function M.getHeaders()
-  local headers = {
-    "NAME",
-    "STATUS",
-    "ROLES",
-    "AGE",
-    "VERSION",
-  }
-
-  return headers
+function M.NodeDesc(node)
+  ResourceBuilder:new("desc", { "describe", "node", node })
+    :fetch()
+    :splitData()
+    :displayFloat("k8s_node_desc", "node_desc", "yaml")
 end
 
 return M
